requirements.txt
opencv-python~=4.11.0.86
pytesseract~=0.3.13
pyserial~=3.5
ultralytics~=8.3.121
pandas~=2.2.3
Flask~=3.1.1
Flask-SocketIO~=5.5.1